auto 0 9 81

+++ Makefile
@@ -46,12 +46,12 @@ setup: libxml2 html_when
 	$(MAKE) -C html_when setup
 
 libxml2: ./html_when/libxml2
-	ln -rs $< $@
+	[[ -l $@ ]] || ln -rs $< $@
 ./html_when/libxml2: html_when
 
 define SYNC
 	if [[ ! -d $1 ]]; then \
-		git clone --recurse-submodules $2 pending-$1 && \
+		git clone $2 pending-$1 && \
 		mv pending-$1 $1 ; \
 	else \
 		cd $1 && git pull; \
